pyproject.toml
[project]
name = "github-follow-contributions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx[socks]>=0.28.1",
    "openai>=2.2.0",
    "pygithub>=2.5.0",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scalene>=1.5.55",
    "tqdm>=4.67.1",
]
